Move source files to SimCore subdirectory.
