+++ src/SimCore/AuxInfoReader.cxx
