+++ src/SimCore/EcalHitIO.cxx
@@ -6,8 +6,8 @@
 #include <map>
 
 // LDMX
-#include "SimCore/SimCalorimeterHit.h"
-#include "SimCore/SimParticle.h"
+#include "SimCore/Event/SimCalorimeterHit.h"
+#include "SimCore/Event/SimParticle.h"
 
 namespace ldmx {
