+++ src/SimCore/G4Session.cxx
