+++ src/SimCore/LHEParticle.cxx
