+++ src/SimCore/LHEReader.cxx
