+++ src/SimCore/MagneticFieldMap3D.cxx
