+++ src/SimCore/TrigScintSD.cxx
