+++ src/SimCore/UserEventInformation.cxx
